Implement analysis_attack()
It has to be used only after the zeroth analysis of its variations.
Moreover I have to be careful to check if it works if
self.available_moves is empty.

The idea is quite easy. I just look the variations,I study each
of them as defence, one study_depth less, and then choose the option
that grants the best advantage.