test: test all the process from parsepgn to light analysis
I parsed, enriched, trimmed, lightened and analyzed many games from the
elite_pgn file. For every step, I recorded the %time and I saved the
result for later use in jupyter notebooks.

The analysis algorithm flows without mistakes and the results seem
reasonable, however the algorithm flows unexpectedly fast, so I will
need to analyze in more detail some example to see if there is anything
to debug. Proably not, but better safe than sorry.

So far the main area of improvement is the parsing of games. I am using
the library pychess but apparently it is a bit of an elephant. I will
still use to visualize the boards, but for the parsing I think I will
move to some c library wrapped with python.